fixed some error in profile.
