+++ src/main/java/UIController/CourseMatchUIControl.java
@@ -4,28 +4,25 @@ import Entities.User;
 import Entities.Student;
 import UseCases.CourseDataManager;
 import UseCases.CourseMatchManager;
-import UseCases.CourseMatchManager2;
 import UseCases.UserDataManager;
-import org.checkerframework.checker.units.qual.A;
 
 import java.io.IOException;
 import java.util.ArrayList;
 
-public class MatchUIControl {
+public class CourseMatchUIControl {
     private CourseMatchManager courseMatchManager;
-    private CourseMatchManager2 courseMatchManager2;
 
     public Student self;
     private ArrayList<Student> matches = new ArrayList();
-    public MatchUIControl(User self, CourseDataManager courseDatabase, UserDataManager userDatabase){
+    public CourseMatchUIControl(User self, CourseDataManager courseDatabase, UserDataManager userDatabase){
         this.self = (Student)self;
-        this.courseMatchManager2 = new CourseMatchManager2(courseDatabase, userDatabase);
+        this.courseMatchManager = new CourseMatchManager(courseDatabase, userDatabase);
     }
 
 
     public ArrayList<Student> getMatches(int min_numCommon) throws IOException {
         this.matches = new ArrayList();
-        this.matches = this.courseMatchManager2.getTopSameSessionStudents(this.self, min_numCommon);
+        this.matches = this.courseMatchManager.getTopSameSessionStudents(this.self, min_numCommon);
         return this.matches;
     }
 
@@ -34,7 +31,7 @@ public class MatchUIControl {
     }
 
     private ArrayList<Student> filterByLabel(String label){
-        return courseMatchManager2.filterByLabel(this.matches, label);
+        return courseMatchManager.filterByLabel(this.matches, label);
     }
 
     private boolean noMatches(){
