+++ src/main/java/UseCases/CourseMatchManager.java
@@ -1,10 +1,11 @@
 package UseCases;
 
+import Entities.Label;
 import Entities.Student;
-import org.checkerframework.checker.units.qual.A;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 
 public class CourseMatchManager {
@@ -17,85 +18,135 @@ public class CourseMatchManager {
         this.ub = userDatabase;
     }
 
+    private int maxSameCourses = 17;
+
+    private int resultLimit = 5;
+
+    private HashMap<String, Integer> matchCount = new HashMap<>();
+
     /**
      * Helper method, stay private.
-     * @return a list of students that has same session as "student". Duplicates expected.
      */
-    private ArrayList<Student> getSameSessionList(Student student){
-        ArrayList<Student> sameSessionStudents = new ArrayList<>();
+    private void updateMatchCount(Student student){
+        matchCount = new HashMap<>();
         ArrayList<String> enrolledCourse = student.getEnrolledCourseCodes();
-        System.out.println(enrolledCourse);
         for(String course: enrolledCourse){
             try {
-                System.out.println(course);
                 ArrayList<String> studentIDs = cb.getCourse(course, "LEC").getEnrolledIDList();
-                ArrayList<Student> students = new ArrayList<>();
-                for (String s: studentIDs){
-                    students.add((Student) ub.getUserByID(s));
+
+                for(String id : studentIDs){
+                    if(!id.equals(student.getUserID())){
+                        if(matchCount.containsKey(id)){
+                            int currentCount = matchCount.get(id);
+                            currentCount ++;
+                            matchCount.put(id, currentCount);
+                        }else{
+                            matchCount.put(id, 1);
+                        }
+                    }
+
                 }
-                sameSessionStudents.addAll(students);
+
+
+
             } catch (IOException e) {
                 System.out.println("CourseMatchManager.java getSameSessionList error probably related to cb.getLecCourse()");
                 throw new RuntimeException(e);
             }
         }
-        return sameSessionStudents;
     }
 
-    /**
-     * Helper method, stay private.
-     */
-    private HashMap<String, Integer> getSameSessionMap(Student student){
-        ArrayList<Student> sameSessionStudents = getSameSessionList(student);
-        HashMap<String, Integer> potentialStudents = new HashMap<>();
-        for(Student s: sameSessionStudents){
-            Integer rank;
-            if(potentialStudents.containsKey(s.getUserID())){
-                rank = potentialStudents.get(s.getUserID());
-                potentialStudents.replace(s.getUserID(), rank, rank + 1);
-            }
-            else {
-                rank = 1;
-                potentialStudents.put(s.getUserID(), rank);
-            }
+    private ArrayList<String>[] getMatchedArray(){
+        String[] keys = this.matchCount.keySet().toArray(new String[0]);
+
+        ArrayList<String>[] matchedList = new ArrayList[this.maxSameCourses];
+
+//      intialize matched list to be an array of empty arrayLists...
+        for(int i = 0; i < matchedList.length; i++){
+            matchedList[i] = new ArrayList<>();
         }
-        return potentialStudents;
+
+        for(String id: keys){
+            int count = this.matchCount.get(id);
+            matchedList[count].add(id);
+        }
+        return matchedList;
     }
 
-    /**
-     * get an arraylist of student that are enrolled in "numOfCommon" same sessions as "student".
-     * @param numOfCommon number of common sessions
-     * @return an arraylist of student
-     */
-    public ArrayList<String> getSameSessionStudents(Student student, int numOfCommon){
-        ArrayList<String> sameSessionStuByNum = new ArrayList<>();
-        HashMap<String, Integer> potentialStudents = getSameSessionMap(student);
-        for(String s: potentialStudents.keySet()){
-            if(potentialStudents.get(s).equals(numOfCommon)){
-                sameSessionStuByNum.add(s);
+
+    private int maxCommon(ArrayList<String>[] matchedArray){
+
+        int latestEmpty = -1;
+
+        for(int i =1; i<this.maxSameCourses; i++){
+            if(matchedArray[i].isEmpty()){
+                latestEmpty = i;
             }
         }
-        return sameSessionStuByNum;
+
+        if(latestEmpty < 0){
+            return this.maxSameCourses;
+        }else{
+            return latestEmpty - 1;
+        }
     }
 
-    /**
-     * get an arraylist of the top "topNum" of students that are enrolled in "numOfCommon" same sessions as "student".
-     * @param numOfCommon number of common sessions
-     * @param topNum number of student of the returned arraylist.
-     * @return an arraylist of student
-     */
-    public ArrayList<String> getTopSameSessionStudents(Student student, int numOfCommon, int topNum){
-        ArrayList<String> sameSessionStu = getSameSessionStudents(student, numOfCommon);
-        ArrayList<String> topSameSessionStu = new ArrayList<>();
-        System.out.println(sameSessionStu.size());
-        if(topNum < sameSessionStu.size()){
-            for(int i = 0; i < topNum; i++){
-                topSameSessionStu.add(sameSessionStu.get(i));
+
+    public ArrayList<Student> getTopSameSessionStudents(Student student, int minNumOfCommon) throws IOException {
+
+        updateMatchCount(student);
+        ArrayList<String>[] matchedList = this.getMatchedArray();
+        int maxCommonCount = maxCommon(matchedList);
+
+        ArrayList<String>[] targetMatchedList = Arrays.copyOfRange(matchedList, minNumOfCommon, maxCommonCount + 1);
+
+        ArrayList<String> ids = new ArrayList<>();
+        for(int i = 0; i<targetMatchedList.length; i++){
+            for(int j = 0; j < targetMatchedList[i].size(); j++){
+                if(!targetMatchedList[i].get(j).isEmpty()){
+                    ids.add(targetMatchedList[i].get(j));
+                }
+
             }
         }
-        for(String s: topSameSessionStu){
-            System.out.println(s);
+
+        ArrayList<Student> students = new ArrayList<>();
+        int idCounts = ids.size();
+
+        if(idCounts > resultLimit){
+            idCounts = resultLimit;
         }
-        return topSameSessionStu;
+
+        System.out.println(idCounts);
+        System.out.println(ids);
+
+        for(int i=0; i<idCounts; i++){
+
+            System.out.println("id = " + ids.get(i));
+
+            if(!(ids.get(i).equals(student.getUserID()))){
+
+                Student stu = (Student) this.ub.getUserByID(ids.get(i));
+                students.add(stu);
+            }
+        }
+
+        return students;
+
+    }
+
+    public ArrayList<Student> filterByLabel(ArrayList<Student> matches, String label){
+        ArrayList<Student> filtered = new ArrayList<>();
+        for(Student s: matches){
+            ArrayList<String> labelNames = new ArrayList<>();
+
+            for(Label l: s.getLabels()){
+                labelNames.add(l.getName());
+            }
+            if(labelNames.contains(label)){
+                filtered.add(s);
+            }
+        }
+        return filtered;
     }
 }
